add plot A vs S
the previous plot on the subplot seems to be correct. the difference to the lecture plot is the scaling, that makes it look different. the plot added here makes it more similar. the other difference is that the map extent, and possibly the original DEM, may not be exactly the same.